+++ Behaviors/Phantom.cs
@@ -23,14 +23,30 @@ namespace SilkenSisters.Behaviors
 
         private async Task Setup()
         {
-            SilkenSisters.Log.LogMessage($"[PhantomBoss.Setup] Started setting phantom boss up");
-            gameObject.transform.SetPositionX(77.1797f);
-            getComponents();
-            triggerLace();
-            skipCutscene();
-            listenForLaceDead();
-            prepareExitMemoryEffect();
-            SilkenSisters.Log.LogMessage($"[PhantomBoss.Setup] Finished setting phantom boss up");
+            try
+            {
+                SilkenSisters.Log.LogMessage($"[PhantomBoss.Setup] Started setting phantom boss up");
+                gameObject.transform.SetPositionX(77.1797f);
+                getComponents();
+                triggerLace();
+                listenForLaceDead();
+
+                if (SilkenSisters.isMemory()) { 
+                    skipCutscene();
+                    prepareExitMemoryEffect();
+                }
+
+                prepareSync();
+
+                InvokeMethod constrainHornet = new InvokeMethod(endHornetConstrain);
+                _control.AddAction("Final Parry", constrainHornet);
+
+                SilkenSisters.Log.LogMessage($"[PhantomBoss.Setup] Finished setting phantom boss up");
+            }
+            catch (Exception e)
+            {
+                SilkenSisters.Log.LogError($"{e} {e.Message}");
+            }
         }
 
         private void getComponents()
@@ -58,7 +74,7 @@ namespace SilkenSisters.Behaviors
             target_boss.gameObject = SilkenSisters.plugin.laceBossFSMOwner;
             target_boss.target = FsmEventTarget.EventTarget.GameObject;
             lace_boss_start.eventTarget = target_boss;
-
+            
             _control.AddAction("To Idle", lace_boss_start);
         }
 
@@ -90,7 +106,7 @@ namespace SilkenSisters.Behaviors
             laceBossVar.SetName("LaceBoss2");
 
             FindGameObject laceObject = new FindGameObject();
-            laceObject.objectName = $"{SilkenSisters.plugin.lace2BossInstance.name}";
+            laceObject.objectName = $"{SilkenSisters.plugin.laceBossInstance.name}";
             laceObject.store = laceBossVar;
             laceObject.withTag = "Untagged";
 
@@ -245,9 +261,23 @@ namespace SilkenSisters.Behaviors
 
             _control.InsertAction("Collapse", disablePhantom, 0);
             _control.InsertAction("Collapse", world_black_thread, 0);
+        }
 
-            InvokeMethod constrainHornet = new InvokeMethod(endHornetConstrain);
-            _control.AddAction("Final Parry", constrainHornet);
+        private void prepareSync()
+        {
+            if (SilkenSisters.syncedFight.Value)
+            {
+                _control.AddState("SilkenSync");
+
+                if (FsmEvent.GetFsmEvent("PHANTOM_SYNC") == null)
+                {
+                    FsmEvent phantomSync = new FsmEvent("PHANTOM_SYNC");
+                    FsmEvent.AddFsmEvent(phantomSync);
+                }
+
+                _control.ChangeTransition("Phase?", "FINISHED", "SilkenSync");
+                _control.AddTransition("SilkenSync", "PHANTOM_SYNC", "Range Check");
+            }
         }
 
         private void endHornetConstrain()
@@ -257,4 +287,111 @@ namespace SilkenSisters.Behaviors
         }
 
     }
+
+    internal class PhantomScene : MonoBehaviour
+    {
+
+        private PlayMakerFSM _control;
+
+        private void Awake()
+        {
+            Setup();
+        }
+
+        private async Task Setup()
+        {
+            getComponents();
+            disableAreaDetection();
+            editFSMEvents();
+            editBossTitle();
+            setupHornetControl();
+        }
+
+        private void getComponents()
+        {
+            _control = gameObject.GetFsmPreprocessed("Control");
+        }
+
+        private void disableAreaDetection()
+        {
+            ((PlayMakerUnity2DProxy)GetComponent(typeof(PlayMakerUnity2DProxy))).enabled = false;
+            ((BoxCollider2D)GetComponent(typeof(BoxCollider2D))).enabled = false;
+        }
+
+        private void editFSMEvents()
+        {
+            SilkenSisters.Log.LogMessage($"[PhantomBoss.editFSMEvents] Trigger lace jump");
+            SendEventByName lace_jump_event = new SendEventByName();
+            lace_jump_event.sendEvent = "ENTER";
+            lace_jump_event.delay = 0;
+
+            FsmEventTarget target = new FsmEventTarget();
+            target.gameObject = SilkenSisters.plugin.laceNPCFSMOwner;
+            target.target = FsmEventTarget.EventTarget.GameObject;
+
+            lace_jump_event.eventTarget = target;
+
+            _control.AddAction("Organ Hit", lace_jump_event);
+
+            FaceObjectV2 hornetFaceEnemies = new FaceObjectV2();
+            hornetFaceEnemies.objectA = SilkenSisters.hornetFSMOwner;
+            hornetFaceEnemies.objectB = SilkenSisters.plugin.laceBossInstance;
+            hornetFaceEnemies.spriteFacesRight = false;
+            hornetFaceEnemies.playNewAnimation = false;
+            hornetFaceEnemies.newAnimationClip = "";
+            hornetFaceEnemies.resetFrame = false;
+            hornetFaceEnemies.everyFrame = false;
+            hornetFaceEnemies.pauseBetweenTurns = 0.1f;
+            _control.AddAction("BG Fog", hornetFaceEnemies);
+
+
+            Tk2dPlayAnimation hornetChall = new Tk2dPlayAnimation();
+            hornetChall.gameObject = SilkenSisters.hornetFSMOwner;
+            hornetChall.clipName = "Challenge Talk Start";
+            hornetChall.animLibName = "";
+            _control.AddAction("BG Fog", hornetChall);
+
+        }
+
+        private void setupHornetControl()
+        {
+            SilkenSisters.Log.LogMessage("[PhantomBoss.setupHornetControl] Setting actions to give back hornet control");
+            SendMessage message_control_idle = new SendMessage();
+            FunctionCall fc_control_idle = new FunctionCall();
+            fc_control_idle.FunctionName = "StartControlToIdle";
+
+            message_control_idle.functionCall = fc_control_idle;
+            message_control_idle.gameObject = SilkenSisters.hornetFSMOwner;
+            message_control_idle.options = SendMessageOptions.DontRequireReceiver;
+
+            SendMessage message_control_regain = new SendMessage();
+            FunctionCall fc_control_regain = new FunctionCall();
+            fc_control_regain.FunctionName = "RegainControl";
+            message_control_regain.functionCall = fc_control_regain;
+            message_control_regain.gameObject = SilkenSisters.hornetFSMOwner;
+            message_control_regain.options = SendMessageOptions.DontRequireReceiver;
+
+            HutongGames.PlayMaker.Actions.SetPlayerDataBool enablePause = new HutongGames.PlayMaker.Actions.SetPlayerDataBool();
+            enablePause.boolName = "disablePause";
+            enablePause.value = false;
+
+
+            Tk2dPlayAnimation hornetChallEnd = new Tk2dPlayAnimation();
+            hornetChallEnd.gameObject = SilkenSisters.hornetFSMOwner;
+            hornetChallEnd.clipName = "Challenge Talk End";
+            hornetChallEnd.animLibName = "";
+
+            _control.AddAction("Start Battle", hornetChallEnd);
+            _control.AddAction("Start Battle", message_control_regain);
+            _control.AddAction("Start Battle", message_control_idle);
+            _control.AddAction("Start Battle", enablePause);
+        }
+
+        private void editBossTitle()
+        {
+            _control.GetAction<DisplayBossTitle>("Start Battle", 3).bossTitle = "SILKEN_SISTERS";
+            SilkenSisters.Log.LogInfo($"[PhantomBoss.editBossTitle] NewTitleBase:{_control.GetAction<DisplayBossTitle>("Start Battle", 3).bossTitle}");
+        }
+    }
+
 }